% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSS.R
\name{calcSS}
\alias{calcSS}
\title{Wrapper for pmsampsize with R-squared estimation and allowing for imprecision in the estimate of rate and mean follow-up.}
\usage{
calcSS(rate, time, parameters, meanfollowup, imprecision = 0.2)
}
\arguments{
\item{rate}{Event rate at time}

\item{time}{Time point to show performance at}

\item{parameters}{Number of parameters in the model}

\item{meanfollowup}{The mean follow-up time}

\item{imprecision}{Allowance for imprecision in the rate and mean follow-up}
}
\value{
GGPlot object
}
\description{
Wrapper for pmsampsize with R-squared estimation and allowing for imprecision in the estimate of rate and mean follow-up.
}
\examples{
calcSS(
  rate = 0.1,
  time = 3,
  parameters = 5,
  meanfollowup = 4
)
}
