% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survPlots.R
\name{survPlots}
\alias{survPlots}
\title{Plots for survival models}
\usage{
survPlots(
  model,
  time,
  df,
  eventVar = "Event",
  timeVar = "Time",
  plotType = "ROC",
  xlim = c(0, 1),
  ylim = c(0, 1),
  colour = "black",
  harm = 0
)
}
\arguments{
\item{model}{CoxPH model object}

\item{time}{Time point to show performance at}

\item{df}{Data contained in data frame}

\item{eventVar}{Name of the event variable in df}

\item{timeVar}{Name of the time variable in df}

\item{plotType}{One of "ROC", "Calibration", or "Decision"}

\item{xlim}{Limit of the x-axis for "Calibration" or "Decision"}

\item{ylim}{Limit of the y-axis for "Calibration"}

\item{colour}{Colour of the plot line}

\item{harm}{Harm of the model for the decision curve}
}
\value{
GGPlot object
}
\description{
Plots for survival models
}
\examples{

# Example using Rotterdam breast cancer data
library(survival)
mod <- coxph(Surv(rtime, recur) ~ age+meno+grade+size, data = rotterdam)
head(rotterdam)
survPlots(
  model = mod,
  time = 2000,
  df = rotterdam,
  eventVar = "recur",
  timeVar = "rtime",
  plotType="ROC")

survPlots(
  model = mod,
  time = 2000,
  df = rotterdam,
  eventVar = "recur",
  timeVar = "rtime",
  plotType="Calibration")

survPlots(
  model = mod,
  time = 2000,
  df = rotterdam,
  eventVar = "recur",
  timeVar = "rtime",
  plotType="Decision")

survPlots(
  model = mod,
  time = 2000,
  df = rotterdam,
  eventVar = "recur",
  timeVar = "rtime",
  plotType="All")
}
